scraper/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
asyncpg==0.30.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
iso8601==2.1.0
pydantic==2.11.7
pydantic_core==2.33.2
pypika-tortoise==0.6.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
soupsieve==2.7
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
